refactor: centralize department constants
