+++ src/constants/departments.ts
@@ -1,7 +1,9 @@
-export const departments = [
+export type Department = { id: string; name: string };
+
+export const departments: Department[] = [
   { id: 'recepcao', name: 'Recepção' },
   { id: 'enfermagem', name: 'Enfermagem' },
   { id: 'medicos', name: 'Médicos' },
   { id: 'limpeza', name: 'Limpeza' },
   { id: 'alimentacao', name: 'Alimentação' }
-]; 
+];
